# feat: optimize bundle size with tsup bundler