--- a/tests/no-hardcoded-jsx-text.test.js
+++ b/tests/no-hardcoded-jsx-text.test.js
@@ -2,10 +2,9 @@
 const { ESLintUtils } = require("@typescript-eslint/utils");
 const path = require("path");
 
-// Load the built rule from lib/
-const rule = require(
-  path.resolve(__dirname, "../lib/no-hardcoded-jsx-text.js"),
-).default;
+// Load the built rule from bundled lib/
+const plugin = require(path.resolve(__dirname, "../lib/index.js"));
+const rule = plugin.rules["no-hardcoded-jsx-text"];
 
 const ruleTester = new ESLintUtils.RuleTester({
   parser: require.resolve("@typescript-eslint/parser"),